# api/openapi.yaml
openapi: 3.0.3
info:
  title: Attachra REST API
  version: "1.0.0"
  description: |
    Attachra's admin/automation REST API (US-8.1, epic E8). This document is
    the **source of truth** for the API contract (ADR / API-first principle):
    the HTTP handlers implementing each operation below (T-8.1.2..T-8.1.7)
    must match it, not the other way around.

    ## Scope

    This spec covers only the versioned, token-authenticated admin API under
    `/api/v1`. Three other HTTP surfaces already implemented by
    `internal/adapters/http` are intentionally **out of scope** here, and are
    listed for completeness rather than documented as paths below:

    - `GET /p/{token}` and `POST /p/{token}/d/{ref}` — the anonymous
      recipient-facing package/download pages (US-6.2). These are not JSON
      resources, carry no Bearer auth, and are documented by their own
      package (`internal/adapters/http/handler.go`,
      `docs/architecture/package-page-decision.md`).
    - `GET /metrics` — Prometheus text exposition format, not JSON; described
      by the metrics library itself, not by an OpenAPI document.
    - `GET /healthz` and `GET /readyz` — liveness/readiness probes
      (`internal/adapters/http/health.go`). Neither is mounted under
      `/api/v1` (`internal/adapters/http/server.go`), so including them as
      `paths` here — which inherit this document's `servers: [{url:
      /api/v1}]` — would misrepresent their real mount point as a contract
...

    ## Versioning

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /attachments:
    get:
      operationId: listAttachments
      summary: List replaced attachments
      tags: [Attachments]
      x-required-role: [admin, viewer]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: message_id
          in: query
          schema: { type: string }
        - name: filename
          in: query
          description: Case-insensitive wildcard pattern matched against the sanitized attachment file name. Supports `*` (any run) and `?` (single character) only — unlike the policy format's `when.attachment.filename`, character classes such as `[a-z]` are not supported and match literally.
          schema: { type: string }
        - name: mime_type
          in: query
          description: Case-insensitive wildcard pattern (`*` and `?` only, no character classes) matched against the *detected* (magic-byte) type, e.g. `application/pdf`, `image/*`.
          schema: { type: string }
        - name: min_size
          in: query
          description: Inclusive lower bound on `size`, in bytes.
          schema: { type: integer, format: int64, minimum: 0 }
        - name: max_size
          in: query
          description: Inclusive upper bound on `size`, in bytes.
          schema: { type: integer, format: int64, minimum: 0 }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: A page of attachments.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttachmentList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /attachments/{attachmentId}:
    get:
      operationId: getAttachment
      summary: Get a single attachment by ID
      tags: [Attachments]
      x-required-role: [admin, viewer]
      parameters:
        - $ref: '#/components/parameters/AttachmentIdPath'
      responses:
        '200':
          description: The attachment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Attachment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        An Attachra API token secret, shown once at creation time
        (`POST /api-tokens`). Never accepted as a query parameter
        (SR-130-2); compared server-side in constant time against the
        stored SHA-256 hash.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous page's `next_cursor`. Omit for the first page.
      schema: { type: string }
...
    AttachmentIdPath:
      name: attachmentId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request (bad query parameter, invalid JSON body, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: bad_request, message: "invalid value for query parameter \"limit\"" } }
    Unauthorized:
      description: Missing, malformed or invalid Bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: "missing or invalid bearer token" } }
    Forbidden:
      description: The token's role is not permitted to call this operation (see `x-required-role`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: forbidden, message: "this token's role is not permitted to perform this action" } }
    NotFound:
      description: No resource exists with the given identifier.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: not_found, message: "no resource found for the given identifier" } }
...
    TooManyRequests:
      description: Rate limit exceeded (SR-130-5).
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: rate_limited, message: "too many requests, slow down" } }
    InternalError:
      description: An unexpected server-side failure. The response never includes internal detail (no stack trace, driver error text, or file path — SR-130-1).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: internal, message: "an internal error occurred" } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error identifier.
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - held
                - invalid_policy
                - validation_failed
                - payload_too_large
                - rate_limited
                - gone
                - internal
            message:
              type: string
              description: Safe, human-readable message. Never includes internal detail (SR-130-1).
            details:
              type: array
              description: Structured, per-field issues. Populated only by policy validate/reload failures.
              items: { $ref: '#/components/schemas/ValidationIssue' }
...
    Attachment:
      type: object
      required: [id, message_id, part_ref, filename, declared_type, detected_type, size, created_at]
      properties:
        id: { type: string }
        message_id: { type: string }
        part_ref: { type: string, description: "store.Attachment.PartRef — the MIME part's position within the message." }
        filename: { type: string, description: Decoded and sanitized attachment file name. }
        declared_type: { type: string, description: "The Content-Type the message declared for this part." }
        detected_type: { type: string, description: The real content type detected from magic bytes. }
        size: { type: integer, format: int64, description: Size in bytes. }
        retain_until:
          type: string
          format: date-time
          nullable: true
          description: Storage retention deadline (US-5.3). Null for legacy pre-ATR-178 rows.
        created_at: { type: string, format: date-time }
      description: >-
        Deliberately excludes `storage_key`: the object store key is an
        opaque, unguessable identifier that is never exposed outside the
        storage driver boundary (SR-121-3) — API consumers address bytes
        only via a link, never directly.
    AttachmentList:
      allOf:
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Attachment' }
        - $ref: '#/components/schemas/PageMeta'
...
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path: { type: string, description: "JSON-pointer-like location, e.g. \"rules[2].then\" or \"default\" (policy.ValidationError.Path)." }
        rule_name: { type: string, description: "Name of the offending rule, when known." }
        message: { type: string }
...
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null if this is the last page.